+++ src/languages.rs
@@ -1,0 +1,11 @@
+//! Defines an enum to specify available languages from this library
+
+/// List of the available languages provided by this library.
+pub enum Language {
+    /// English (generic)
+    En,
+    /// American English
+    EnUs,
+}
+
+pub mod en_us;